Improve layer structure
