+++ lib/model/neuralnetwork.js
@@ -1,7 +1,7 @@
 import Matrix from '../util/matrix.js'
 import Tensor from '../util/tensor.js'
-import { Layer, LossLayer } from './nns/layer/index.js'
-export { Layer } from './nns/layer/index.js'
+import Layer, { LossLayer } from './nns/layer/base.js'
+export { default as Layer } from './nns/layer/base.js'
 
 import ComputationalGraph from './nns/graph.js'
 export { default as ComputationalGraph } from './nns/graph.js'
@@ -13,6 +13,10 @@ import OutputLayer from './nns/layer/output.js'
 
 let ONNXImporter = null
 
+/**
+ * @ignore
+ * @typedef {import("./nns/graph").LayerObject} LayerObject
+ */
 /**
  * Exception for neuralnetwork class
  */
@@ -35,7 +39,7 @@ export default class NeuralNetwork {
 	/**
 	 * Returns neuralnetwork.
 	 *
-	 * @param {Object<string, *>[]} layers Network layers
+	 * @param {LayerObject[]} layers Network layers
 	 * @param {string} [loss] Loss name
 	 * @param {'sgd' | 'adam' | 'momentum' | 'rmsprop'} [optimizer=sgd] Optimizer of the network
 	 * @returns {NeuralNetwork} Created Neuralnetwork
@@ -128,7 +132,7 @@ export default class NeuralNetwork {
 	/**
 	 * Returns object representation.
 	 *
-	 * @returns {object} Object represented this neuralnetwork
+	 * @returns {LayerObject[]} Object represented this neuralnetwork
 	 */
 	toObject() {
 		return this._graph.toObject()
