+++ lib/model/nns/graph.js
@@ -1,10 +1,10 @@
 import Matrix, { MatrixException } from '../../util/matrix.js'
 import { NeuralnetworkException } from '../neuralnetwork.js'
-import { Layer, LossLayer } from './layer/index.js'
-import InputLayer from './layer/input.js'
-import OutputLayer from './layer/output.js'
+import Layer, { LossLayer } from './layer/base.js'
+import { InputLayer, OutputLayer } from './layer/index.js'
 
 /**
+ * @typedef {import("./layer/index").PlainLayerObject & {input?: string | string[], name?: string}} LayerObject
  * @typedef {object} Node
  * @property {Layer} layer Layer
  * @property {string} name Name of the node
@@ -42,7 +42,7 @@ export default class ComputationalGraph {
 	/**
 	 * Returns Graph.
 	 *
-	 * @param {Object<string, *>[]} nodes Array of object represented a graph
+	 * @param {LayerObject[]} nodes Array of object represented a graph
 	 * @returns {ComputationalGraph} Graph
 	 */
 	static fromObject(nodes) {
@@ -60,7 +60,7 @@ export default class ComputationalGraph {
 	/**
 	 * Returns object representation.
 	 *
-	 * @returns {Object<string, *>[]} Object represented this graph
+	 * @returns {LayerObject[]} Object represented this graph
 	 */
 	toObject() {
 		const s = []
