+++ lib/model/nns/layer/index.js
@@ -1,339 +1,244 @@
-import Tensor from '../../../util/tensor.js'
-import Layer from './base.js'
-export { default as Layer, LossLayer, FlowLayer } from './base.js'
+// This file is generated automatically.
+export { default as AddLayer } from './add.js'
+export { default as AdditiveCoupling } from './additive_coupling.js'
+export { default as AdaptivePiecewiseLinearLayer } from './apl.js'
+export { default as ArandaLayer } from './aranda.js'
+export { default as ArgmaxLayer } from './argmax.js'
+export { default as AveragePoolLayer } from './averagepool.js'
+export { default as BatchNormalizationLayer } from './batch_normalization.js'
+export { default as BimodalDerivativeAdaptiveActivationLayer } from './bdaa.js'
+export { default as BendableLinearUnitLayer } from './blu.js'
+export { default as BoundedReLULayer } from './brelu.js'
+export { default as ContinuouslyDifferentiableELULayer } from './celu.js'
+export { default as ClipLayer } from './clip.js'
+export { default as ConcatLayer } from './concat.js'
+export { default as CondLayer } from './cond.js'
+export { default as ConstLayer } from './const.js'
+export { default as ConvLayer } from './conv.js'
+export { default as ConcatenatedReLULayer } from './crelu.js'
+export { default as DetachLayer } from './detach.js'
+export { default as DivLayer } from './div.js'
+export { default as DropoutLayer } from './dropout.js'
+export { default as ElasticELULayer } from './eelu.js'
+export { default as ELULayer } from './elu.js'
+export { default as ElasticReLULayer } from './erelu.js'
+export { default as ESwishLayer } from './eswish.js'
+export { default as FastELULayer } from './felu.js'
+export { default as FlattenLayer } from './flatten.js'
+export { default as FlexibleReLULayer } from './frelu.js'
+export { default as FullyConnected } from './full.js'
+export { default as GaussianLayer } from './gaussian.js'
+export { default as GlobalAveragePoolLayer } from './global_averagepool.js'
+export { default as GlobalLpPoolLayer } from './global_lppool.js'
+export { default as GlobalMaxPoolLayer } from './global_maxpool.js'
+export { default as GRULayer } from './gru.js'
+export { default as HardShrinkLayer } from './hard_shrink.js'
+export { default as HardSigmoidLayer } from './hard_sigmoid.js'
+export { default as HardTanhLayer } from './hard_tanh.js'
+export { default as HexpoLayer } from './hexpo.js'
+export { default as HuberLayer } from './huber.js'
+export { default as IncludeLayer } from './include.js'
+export { default as InputLayer } from './input.js'
+export { default as ImprovedSigmoidLayer } from './isigmoid.js'
+export { default as LeakyReLULayer } from './leaky_relu.js'
+export { default as LessLayer } from './less.js'
+export { default as LinearLayer } from './linear.js'
+export { default as LpPoolLayer } from './lppool.js'
+export { default as LRNLayer } from './lrn.js'
+export { default as LSTMLayer } from './lstm.js'
+export { default as MatmulLayer } from './matmul.js'
+export { default as MaxPoolLayer } from './maxpool.js'
+export { default as MeanLayer } from './mean.js'
+export { default as MultipleParametricELULayer } from './mpelu.js'
+export { default as MSELayer } from './mse.js'
+export { default as MultibinTrainableLinearUnitLayer } from './mtlu.js'
+export { default as MultLayer } from './mult.js'
+export { default as NaturalLogarithmReLULayer } from './nlrelu.js'
+export { default as OnehotLayer } from './onehot.js'
+export { default as OutputLayer } from './output.js'
+export { default as PadeActivationUnitLayer } from './pau.js'
+export { default as ParametricDeformableELULayer } from './pdelu.js'
+export { default as ParametricELULayer } from './pelu.js'
+export { default as PiecewiseLinearUnitLayer } from './plu.js'
+export { default as PowerLayer } from './power.js'
+export { default as ParametricReLULayer } from './prelu.js'
+export { default as ParametricRectifiedExponentialUnitLayer } from './preu.js'
+export { default as ParametricSigmoidFunctionLayer } from './psf.js'
+export { default as PenalizedTanhLayer } from './ptanh.js'
+export { default as ParametricTanhLinearUnitLayer } from './ptelu.js'
+export { default as RandomLayer } from './random.js'
+export { default as RectifiedPowerUnitLayer } from './repu.js'
+export { default as ReshapeLayer } from './reshape.js'
+export { default as ReverseLayer } from './reverse.js'
+export { default as RNNLayer } from './rnn.js'
+export { default as RandomizedReLULayer } from './rrelu.js'
+export { default as RandomTranslationReLULayer } from './rtrelu.js'
+export { default as ScaledELULayer } from './selu.js'
+export { default as SigmoidLayer } from './sigmoid.js'
+export { default as SelfLearnableAFLayer } from './slaf.js'
+export { default as SoftplusLinearUnitLayer } from './slu.js'
+export { default as SoftShrinkLayer } from './soft_shrink.js'
+export { default as SoftargmaxLayer } from './softargmax.js'
+export { default as SoftmaxLayer } from './softmax.js'
+export { default as SoftplusLayer } from './softplus.js'
+export { default as SparseLayer } from './sparse.js'
+export { default as SplitLayer } from './split.js'
+export { default as ShiftedReLULayer } from './srelu.js'
+export { default as SoftRootSignLayer } from './srs.js'
+export { default as ScaledTanhLayer } from './stanh.js'
+export { default as StdLayer } from './std.js'
+export { default as SubLayer } from './sub.js'
+export { default as SumLayer } from './sum.js'
+export { default as SupervisorLayer } from './supervisor.js'
+export { default as SwishLayer } from './swish.js'
+export { default as TrainableAFLayer } from './taf.js'
+export { default as ThresholdedReLULayer } from './thresholded_relu.js'
+export { default as TransposeLayer } from './transpose.js'
+export { default as VariableLayer } from './variable.js'
+export { default as VarLayer } from './variance.js'
 
-const buildLayer = (name, calcFunc, gradFunc) => {
-	class TempLayer extends Layer {
-		calc(x) {
-			this._i = x
-			this._o = x.copy()
-			this._o.map(calcFunc)
-			return this._o
-		}
-
-		grad(bo) {
-			const bi = new Tensor(bo.sizes)
-			for (let i = 0; i < bi.length; i++) {
-				bi.value[i] = bo.value[i] * gradFunc(this._i.value[i], this._o.value[i])
-			}
-			if (bi.dimension === 2) {
-				return bi.toMatrix()
-			}
-			return bi
-		}
-	}
-	Object.defineProperty(TempLayer, 'name', {
-		value: name.split('_').reduce((s, nm) => s + nm[0].toUpperCase() + nm.substring(1).toLowerCase(), '') + 'Layer',
-	})
-	TempLayer.registLayer(name)
-}
-
-const simpleLayers = {
-	abs: {
-		calc: Math.abs,
-		grad: (i, o) => (i < 0 ? -1 : 1),
-	},
-	acos: {
-		calc: Math.acos,
-		grad: (i, o) => -1 / (Math.sqrt(1 - i ** 2) + 1.0e-4),
-	},
-	acosh: {
-		calc: Math.acosh,
-		grad: (i, o) => 1 / (Math.sqrt(i ** 2 - 1) + 1.0e-4),
-	},
-	asin: {
-		calc: Math.asin,
-		grad: (i, o) => 1 / (Math.sqrt(1 - i ** 2) + 1.0e-4),
-	},
-	asinh: {
-		calc: Math.asinh,
-		grad: (i, o) => 1 / Math.sqrt(1 + i ** 2),
-	},
-	atan: {
-		calc: Math.atan,
-		grad: (i, o) => 1 / (1 + i ** 2),
-	},
-	atanh: {
-		calc: Math.atanh,
-		grad: (i, o) => 1 / (1 - i ** 2),
-	},
-	bent_identity: {
-		calc: v => (Math.sqrt(v ** 2 + 1) - 1) / 2 + v,
-		grad: (i, o) => i / (2 * Math.sqrt(i ** 2 + 1)) + 1,
-	},
-	ceil: {
-		calc: Math.ceil,
-		grad: (i, o) => 1,
-	},
-	cloglog: {
-		calc: v => 1 - Math.exp(-Math.exp(v)),
-		grad: (i, o) => Math.exp(i) * (1 - o),
-	},
-	cloglogm: {
-		calc: v => 1 - 2 * Math.exp(-0.7 * Math.exp(v)),
-		grad: (i, o) => 1.4 * Math.exp(i - 0.7 * Math.exp(i)),
-	},
-	cos: {
-		calc: Math.cos,
-		grad: (i, o) => -Math.sin(i),
-	},
-	cosh: {
-		calc: Math.cosh,
-		grad: (i, o) => Math.sinh(i),
-	},
-	elish: {
-		calc: v => (v >= 0 ? v : Math.exp(v) - 1) / (1 + Math.exp(-v)),
-		grad: (i, o) => ((i >= 0 ? 1 : Math.exp(i)) + o * Math.exp(-i)) / (1 + Math.exp(-i)),
-	},
-	elliott: {
-		calc: v => v / (2 * (1 + Math.abs(v))) + 0.5,
-		grad: (i, o) => 1 / (2 * (1 + Math.abs(i)) ** 2),
-	},
-	erf: {
-		calc: v => {
-			const p = 0.3275911
-			const a1 = 0.254829592
-			const a2 = -0.284496736
-			const a3 = 1.421413741
-			const a4 = -1.453152027
-			const a5 = 1.061405429
-
-			const sign = v < 0 ? -1 : 1
-			const t = 1 / (1 + p * Math.abs(v))
-			const erf = 1 - ((((a5 * t + a4) * t + a3) * t + a2) * t + a1) * t * Math.exp(-v * v)
-			return sign * erf
-		},
-		grad: (i, o) => (2 * Math.exp(-(i ** 2))) / Math.sqrt(Math.PI),
-	},
-	exp: {
-		calc: Math.exp,
-		grad: (i, o) => o,
-	},
-	floor: {
-		calc: Math.floor,
-		grad: (i, o) => 1,
-	},
-	geru: {
-		calc: v => {
-			const erf = v => {
-				const p = 0.3275911
-				const a1 = 0.254829592
-				const a2 = -0.284496736
-				const a3 = 1.421413741
-				const a4 = -1.453152027
-				const a5 = 1.061405429
-
-				const sign = v < 0 ? -1 : 1
-				const t = 1 / (1 + p * Math.abs(v))
-				const erf = 1 - ((((a5 * t + a4) * t + a3) * t + a2) * t + a1) * t * Math.exp(-v * v)
-				return sign * erf
-			}
-			return (v * (1 + erf(v / Math.sqrt(2)))) / 2
-		},
-		grad: (i, o) => (i === 0 ? 0 : o / i) + (i * Math.exp(-((i / Math.sqrt(2)) ** 2))) / Math.sqrt(2 * Math.PI),
-	},
-	hard_elish: {
-		calc: v => Math.max(0, Math.min(1, (v + 1) / 2)) * (v >= 0 ? v : Math.exp(v) - 1),
-		grad: (i, o) => (i <= -1 ? 0 : i >= 1 ? (i >= 0 ? 1 : Math.exp(i)) : i >= 0 ? i : ((1 + i) * Math.exp(i)) / 2),
-	},
-	hard_swish: {
-		calc: v => (v <= -3 ? 0 : 3 <= v ? v : (v * (v + 3)) / 6),
-		grad: (i, o) => (i <= -3 ? 0 : 3 <= i ? 1 : (2 * i + 3) / 6),
-	},
-	lisht: {
-		calc: v => v * Math.tanh(v),
-		grad: (i, o) => Math.tanh(i) + i * (1 - Math.tanh(i) ** 2),
-	},
-	log: {
-		calc: Math.log,
-		grad: (i, o) => 1 / i,
-	},
-	loglog: {
-		calc: v => Math.exp(-Math.exp(-v)),
-		grad: (i, o) => Math.exp(-i) * o,
-	},
-	logsigmoid: {
-		calc: v => Math.log(1 / (1 + Math.exp(-v))),
-		grad: (i, o) => 1 / (1 + Math.exp(i)),
-	},
-	mish: {
-		calc: v => v * Math.tanh(Math.log(1 + Math.exp(v))),
-		grad: (i, o) =>
-			(Math.exp(i) * (4 * (i + 1) + 4 * Math.exp(2 * i) + Math.exp(3 * i) + Math.exp(i) * (4 * i + 6))) /
-			(2 * Math.exp(i) + Math.exp(2 * i) + 2) ** 2,
-	},
-	negative: {
-		calc: v => -v,
-		grad: (i, o) => -1,
-	},
-	relu: {
-		calc: v => (v > 0 ? v : 0),
-		grad: (i, o) => (i > 0 ? 1 : 0),
-	},
-	resech: {
-		calc: v => v / Math.cosh(v),
-		grad: (i, o) => 1 / Math.cosh(i) - o * Math.tanh(i),
-	},
-	reu: {
-		calc: v => (v > 0 ? v : v * Math.exp(v)),
-		grad: (i, o) => (i > 0 ? 1 : Math.exp(i) + o),
-	},
-	rootsig: {
-		calc: v => v / (1 + Math.sqrt(1 + v ** 2)),
-		grad: (i, o) => 1 / (Math.sqrt(i ** 2 + 1) + i ** 2 + 1),
-	},
-	round: {
-		calc: Math.round,
-		grad: (i, o) => 1,
-	},
-	sign: {
-		calc: Math.sign,
-		grad: (i, o) => 1,
-	},
-	silu: {
-		calc: v => v / (1 + Math.exp(-v)),
-		grad: (i, o) => 1 / (1 + Math.exp(-i)) + o * (1 - o / i),
-	},
-	sin: {
-		calc: Math.sin,
-		grad: (i, o) => Math.cos(i),
-	},
-	sinh: {
-		calc: Math.sinh,
-		grad: (i, o) => Math.cosh(i),
-	},
-	softsign: {
-		calc: v => v / (1 + Math.abs(v)),
-		grad: (i, o) => 1 / (1 + Math.abs(i)) ** 2,
-	},
-	sqrt: {
-		calc: Math.sqrt,
-		grad: (i, o) => 1 / (2 * o),
-	},
-	square: {
-		calc: v => v ** 2,
-		grad: (i, o) => i * 2,
-	},
-	ssigmoid: {
-		calc: v => 4 / (1 + Math.exp(-v)) - 2,
-		grad: (i, o) => (4 * Math.exp(-i)) / (1 + Math.exp(-i)) ** 2,
-	},
-	tan: {
-		calc: Math.tan,
-		grad: (i, o) => 1 / Math.cos(i) ** 2,
-	},
-	tanh: {
-		calc: Math.tanh,
-		grad: (i, o) => 1 - o ** 2,
-	},
-	tanhexp: {
-		calc: v => v * Math.tanh(Math.exp(v)),
-		grad: (i, o) => Math.tanh(Math.exp(i)) - i * Math.exp(i) * (Math.tanh(Math.exp(i)) ** 2 - 1),
-	},
-	tanhshrink: {
-		calc: v => v - Math.tanh(v),
-		grad: (i, o) => Math.tanh(i) ** 2,
-	},
-}
-
-for (const name of Object.keys(simpleLayers)) {
-	buildLayer(name, simpleLayers[name].calc, simpleLayers[name].grad)
-}
-
-export * from './add.js'
-export * from './additive_coupling.js'
-export * from './apl.js'
-export * from './aranda.js'
-export * from './argmax.js'
-export * from './averagepool.js'
-export * from './batch_normalization.js'
-export * from './bdaa.js'
-export * from './blu.js'
-export * from './brelu.js'
-export * from './celu.js'
-export * from './clip.js'
-export * from './concat.js'
-export * from './cond.js'
-export * from './const.js'
-export * from './conv.js'
-export * from './crelu.js'
-export * from './detach.js'
-export * from './div.js'
-export * from './dropout.js'
-export * from './eelu.js'
-export * from './elu.js'
-export * from './erelu.js'
-export * from './eswish.js'
-export * from './felu.js'
-export * from './flatten.js'
-export * from './frelu.js'
-export * from './full.js'
-export * from './gaussian.js'
-export * from './global_averagepool.js'
-export * from './global_lppool.js'
-export * from './global_maxpool.js'
-export * from './gru.js'
-export * from './hard_shrink.js'
-export * from './hard_sigmoid.js'
-export * from './hard_tanh.js'
-export * from './hexpo.js'
-export * from './huber.js'
-export * from './include.js'
-export * from './input.js'
-export * from './isigmoid.js'
-export * from './leaky_relu.js'
-export * from './less.js'
-export * from './linear.js'
-export * from './lppool.js'
-export * from './lrn.js'
-export * from './lstm.js'
-export * from './matmul.js'
-export * from './maxpool.js'
-export * from './mean.js'
-export * from './mpelu.js'
-export * from './mse.js'
-export * from './mtlu.js'
-export * from './mult.js'
-export * from './nlrelu.js'
-export * from './onehot.js'
-export * from './output.js'
-export * from './pau.js'
-export * from './pdelu.js'
-export * from './pelu.js'
-export * from './plu.js'
-export * from './power.js'
-export * from './prelu.js'
-export * from './preu.js'
-export * from './psf.js'
-export * from './ptanh.js'
-export * from './ptelu.js'
-export * from './random.js'
-export * from './repu.js'
-export * from './reshape.js'
-export * from './reverse.js'
-export * from './rnn.js'
-export * from './rrelu.js'
-export * from './rtrelu.js'
-export * from './selu.js'
-export * from './sigmoid.js'
-export * from './slaf.js'
-export * from './slu.js'
-export * from './soft_shrink.js'
-export * from './softargmax.js'
-export * from './softmax.js'
-export * from './softplus.js'
-export * from './sparse.js'
-export * from './split.js'
-export * from './srelu.js'
-export * from './srs.js'
-export * from './stanh.js'
-export * from './std.js'
-export * from './sub.js'
-export * from './sum.js'
-export * from './supervisor.js'
-export * from './swish.js'
-export * from './taf.js'
-export * from './thresholded_relu.js'
-export * from './transpose.js'
-export * from './variable.js'
-export * from './variance.js'
+/**
+ * @typedef {(
+ * { type: 'abs' } |
+ * { type: 'acos' } |
+ * { type: 'acosh' } |
+ * { type: 'add' } |
+ * { type: 'additive_coupling', d?: number | null, net?: NeuralNetwork | *[] | null } |
+ * { type: 'apl', s?: number, a?: number | number[], b?: number | number[] } |
+ * { type: 'aranda', l?: number } |
+ * { type: 'argmax' } |
+ * { type: 'asin' } |
+ * { type: 'asinh' } |
+ * { type: 'atan' } |
+ * { type: 'atanh' } |
+ * { type: 'average_pool', kernel: number | number[], stride?: number | number[], padding?: number | number[], channel_dim?: number } |
+ * { type: 'batch_normalization', scale?: number, offset?: number } |
+ * { type: 'bdaa', alpha?: number } |
+ * { type: 'bent_identity' } |
+ * { type: 'blu', beta?: number } |
+ * { type: 'brelu', a?: number } |
+ * { type: 'ceil' } |
+ * { type: 'celu', a?: number } |
+ * { type: 'clip', min?: number, max?: number } |
+ * { type: 'cloglog' } |
+ * { type: 'cloglogm' } |
+ * { type: 'concat', axis?: number } |
+ * { type: 'cond' } |
+ * { type: 'const', value: number } |
+ * { type: 'conv', kernel: number | number[], channel?: number, stride?: number | number[], padding?: number | number[], w?: number[][] | Tensor, activation?: string, l2_decay?: number, l1_decay?: number, activation_params?: object, channel_dim?: number } |
+ * { type: 'cos' } |
+ * { type: 'cosh' } |
+ * { type: 'crelu' } |
+ * { type: 'detach' } |
+ * { type: 'div' } |
+ * { type: 'dropout', drop_rate?: number } |
+ * { type: 'eelu', k?: number, alpha?: number, beta?: number } |
+ * { type: 'elish' } |
+ * { type: 'elliott' } |
+ * { type: 'elu', a?: number } |
+ * { type: 'erelu' } |
+ * { type: 'erf' } |
+ * { type: 'eswish', beta?: number } |
+ * { type: 'exp' } |
+ * { type: 'felu', alpha?: number } |
+ * { type: 'flatten' } |
+ * { type: 'floor' } |
+ * { type: 'frelu', b?: number } |
+ * { type: 'full', out_size: number | string, w?: number[][] | Matrix, b?: number[][] | Matrix, activation?: string, l2_decay?: number, l1_decay?: number, activation_params?: object } |
+ * { type: 'gaussian' } |
+ * { type: 'geru' } |
+ * { type: 'global_average_pool', channel_dim?: number } |
+ * { type: 'global_lp_pool', p?: number, channel_dim?: number } |
+ * { type: 'global_max_pool', channel_dim?: number } |
+ * { type: 'gru', size: number, return_sequences?: boolean, w_z?: number[][] | Matrix, w_r?: number[][] | Matrix, w_h?: number[][] | Matrix, u_z?: number[][] | Matrix, u_r?: number[][] | Matrix, u_h?: number[][] | Matrix, b_z?: number[][] | Matrix, b_r?: number[][] | Matrix, b_h?: number[][] | Matrix } |
+ * { type: 'hard_elish' } |
+ * { type: 'hard_shrink', l?: number } |
+ * { type: 'hard_sigmoid', alpha?: number, beta?: number } |
+ * { type: 'hard_swish' } |
+ * { type: 'hard_tanh', v?: number } |
+ * { type: 'hexpo', a?: number, b?: number, c?: number, d?: number } |
+ * { type: 'huber' } |
+ * { type: 'include', net: NeuralNetwork | object[], input_to?: string, train?: boolean } |
+ * { type: 'input', name?: string } |
+ * { type: 'isigmoid', a?: number, alpha?: number } |
+ * { type: 'leaky_relu', a?: number } |
+ * { type: 'less' } |
+ * { type: 'linear' } |
+ * { type: 'lisht' } |
+ * { type: 'log' } |
+ * { type: 'loglog' } |
+ * { type: 'logsigmoid' } |
+ * { type: 'lp_pool', p?: number, kernel: number | number[], stride?: number | number[], padding?: number | number[], channel_dim?: number } |
+ * { type: 'lrn', alpha?: number, beta?: number, k?: number, n: number, channel_dim?: number } |
+ * { type: 'lstm', size: number, return_sequences?: boolean, w_z?: number[][] | Matrix, w_in?: number[][] | Matrix, w_for?: number[][] | Matrix, w_out?: number[][] | Matrix, r_z?: number[][] | Matrix, r_in?: number[][] | Matrix, r_for?: number[][] | Matrix, r_out?: number[][] | Matrix, p_in?: number[][] | Matrix, p_for?: number[][] | Matrix, p_out?: number[][] | Matrix, b_z?: number[][] | Matrix, b_in?: number[][] | Matrix, b_for?: number[][] | Matrix, b_out?: number[][] | Matrix } |
+ * { type: 'matmul' } |
+ * { type: 'max_pool', kernel: number | number[], stride?: number | number[], padding?: number | number[], channel_dim?: number } |
+ * { type: 'mean', axis?: number } |
+ * { type: 'mish' } |
+ * { type: 'mpelu', alpha?: number, beta?: number } |
+ * { type: 'mse' } |
+ * { type: 'mtlu', a?: number | number[], b?: number | number[], c?: number | number[], k?: number } |
+ * { type: 'mult' } |
+ * { type: 'negative' } |
+ * { type: 'nlrelu', beta?: number } |
+ * { type: 'onehot', class_size?: number, values?: number } |
+ * { type: 'output' } |
+ * { type: 'pau', m?: number, n?: number, a?: number | number[], b?: number | number[] } |
+ * { type: 'pdelu', t?: number, alpha?: number } |
+ * { type: 'pelu', a?: number, b?: number } |
+ * { type: 'plu', alpha?: number, c?: number } |
+ * { type: 'power', n: number } |
+ * { type: 'prelu', a?: number } |
+ * { type: 'preu', alpha?: number, beta?: number } |
+ * { type: 'psf', m?: number } |
+ * { type: 'ptanh', a?: number } |
+ * { type: 'ptelu', alpha?: number, beta?: number } |
+ * { type: 'random', size: number | number[] | string, mean?: number, variance?: number } |
+ * { type: 'relu' } |
+ * { type: 'repu', s?: number } |
+ * { type: 'resech' } |
+ * { type: 'reshape', size: number[] | string } |
+ * { type: 'reu' } |
+ * { type: 'reverse', axis?: number } |
+ * { type: 'rnn', size: number, out_size?: number, activation?: string, recurrent_activation?: string, return_sequences?: boolean, w_xh?: number[][] | Matrix, w_hh?: number[][] | Matrix, w_hy?: number[][] | Matrix, b_xh?: number[][] | Matrix, b_hh?: number[][] | Matrix, b_hy?: number[][] | Matrix, activation_params?: object, recurrent_activation_params?: object } |
+ * { type: 'rootsig' } |
+ * { type: 'round' } |
+ * { type: 'rrelu', l?: number, u?: number } |
+ * { type: 'rtrelu' } |
+ * { type: 'selu', a?: number, g?: number } |
+ * { type: 'sigmoid', a?: number } |
+ * { type: 'sign' } |
+ * { type: 'silu' } |
+ * { type: 'sin' } |
+ * { type: 'sinh' } |
+ * { type: 'slaf', n?: number, a?: number | number[] } |
+ * { type: 'slu', alpha?: number, beta?: number, gamma?: number } |
+ * { type: 'soft_shrink', l?: number } |
+ * { type: 'softargmax', beta?: number } |
+ * { type: 'softmax' } |
+ * { type: 'softplus', beta?: number } |
+ * { type: 'softsign' } |
+ * { type: 'sparsity', rho: number, beta: number } |
+ * { type: 'split', axis?: number, size: number | number[] } |
+ * { type: 'sqrt' } |
+ * { type: 'square' } |
+ * { type: 'srelu', d?: number } |
+ * { type: 'srs', alpha?: number, beta?: number } |
+ * { type: 'ssigmoid' } |
+ * { type: 'stanh', a?: number, b?: number } |
+ * { type: 'std', axis?: number } |
+ * { type: 'sub' } |
+ * { type: 'sum', axis?: number } |
+ * { type: 'supervisor' } |
+ * { type: 'swish', beta?: number } |
+ * { type: 'taf', a?: number, b?: number } |
+ * { type: 'tan' } |
+ * { type: 'tanh' } |
+ * { type: 'tanhexp' } |
+ * { type: 'tanhshrink' } |
+ * { type: 'thresholded_relu', a?: number } |
+ * { type: 'transpose', axis: number[] } |
+ * { type: 'variable', size: number[], l2_decay?: number, l1_decay?: number, value?: number[] | number[][] | Tensor } |
+ * { type: 'variance', axis?: number }
+ * )} PlainLayerObject
+ */
